# backend/openapi.yaml
openapi: 3.0.3
info:
  title: QuantDog API
  version: 1.0.0
  description: QuantDog stock market analysis API with technical indicators, sentiment analysis, and AI-powered research capabilities.
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.quantdog.com
    description: Production server
paths:
...
  /api/v1/stocks/{symbol}/strategy:
    post:
      summary: Stock strategy
      description: Get comprehensive strategy analysis for a stock
      parameters:
        - in: path
          name: symbol
          schema:
            type: string
          required: true
          description: Stock symbol
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyRequest'
      responses:
        '200':
          description: Strategy analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
  /api/v1/stocks/{symbol}/monitor:
    get:
      summary: Monitor single stock
      description: Get monitoring snapshot for a single stock
      parameters:
        - in: path
          name: symbol
          schema:
            type: string
          required: true
        - in: query
          name: horizon
          schema:
            type: string
            default: 1d
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Monitoring data retrieved
          content:
            application/json:
              type: object
  /api/v1/stocks/monitor:
    post:
      summary: Monitor multiple stocks
      description: Get monitoring snapshots for multiple stocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitorRequest'
      responses:
        '200':
          description: Monitoring data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitorResponse'
...
components:
  schemas:
...
    EnvelopeSuccess:
      type: object
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          enum: [1]
        msg:
          type: string
        data:
          type: object
...
    StrategyRequest:
      type: object
      properties:
        horizon:
          type: string
          default: 1d
        limit:
          type: integer
          default: 20
    StrategyResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeSuccess'
        - type: object
          properties:
            data:
              type: object
              properties:
                decision:
                  type: string
                  enum: [BUY, SELL, HOLD]
                confidence:
                  type: integer
                combined_score:
                  type: number
                inputs:
                  type: object
                risk_filter:
                  type: object
    MonitorRequest:
      type: object
      required:
        - symbols
      properties:
        symbols:
          type: array
          items:
            type: string
        horizon:
          type: string
          default: 1d
        limit:
          type: integer
          default: 20
    MonitorResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeSuccess'
        - type: object
          properties:
            data:
              type: object
              properties:
                results:
                  type: array
                  items:
                    type: object
                alerts:
                  type: array
                  items:
                    type: object
...